requirements/dev.txt
Django==2.0.13
django-extensions==2.1.6
django-rest==0.0.1
djangorestframework==3.9.2
ipdb==0.11
ipython==6.5.0
requests==2.21.0
